+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ASB + CONTROLLER: Add a `monero_seed` command to the controller shell. You can use it to export the seed and restore height of the internal Monero wallet. You can use those to import the wallet into a wallet software of your own choosing.
 - GUI: You can now change the Monero Node without having to restart.
+- GUI: You can now export the seed phrase of the Monero wallet.
 
 ## [3.0.0-beta.8] - 2025-08-10